# internal/handler/userHandler.go
package handler

import (
	"encoding/json"
	"medivault-service/internal/services"
	"medivault-service/pkg/utils"
	"net/http"
)

type UserHandler struct {
	Service services.UserService
}

// CreateUser creates a new user
//
//	@Summary		Create a new user
//	@Description	Register a new user in the system
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			user	body		User	true	"User data"
//	@Success		201		{object}	User
//	@Failure		400		{object}	utils.ErrorResponse
//	@Failure		500		{object}	utils.ErrorResponse
//	@Router			/users/create [post]
func (u *UserHandler) CreateUser(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "CreateUser functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetUserByID retrieves a user by ID
//
//	@Summary		Get user by ID
//	@Description	Retrieve a user by ID
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"User ID"
//	@Success		200	{object}	User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		404	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/users/{id} [get]
func (u *UserHandler) GetUserByID(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetUserByID functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetAllUsers retrieves all users
//
//	@Summary		Get all users
//	@Description	Retrieve all users
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	[]User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/users [get]
func (u *UserHandler) GetAllUsers(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetAllUsers functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetUserByEmail retrieves a user by email
//
//	@Summary		Get user by email
//	@Description	Retrieve a user by email
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			email	path		string	true	"User email"
//	@Success		200		{object}	User
//	@Failure		400		{object}	utils.ErrorResponse
//	@Failure		404		{object}	utils.ErrorResponse
//	@Failure		500		{object}	utils.ErrorResponse
//	@Router			/users/email/{email} [get]
func (u *UserHandler) GetUserByEmail(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetUserByEmail functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// UpdateUser updates a user
//
//	@Summary		Update a user
//	@Description	Update a user in the system
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id		path		int		true	"User ID"
//	@Param			user	body		User	true	"User data"
//	@Success		200		{object}	User
//	@Failure		400		{object}	utils.ErrorResponse
//	@Failure		404		{object}	utils.ErrorResponse
//	@Failure		500		{object}	utils.ErrorResponse
//	@Router			/users/{id} [put]
func (u *UserHandler) UpdateUser(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "UpdateUser functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// DeleteUser deletes a user
//
//	@Summary		Delete a user
//	@Description	Delete a user in the system
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id	path		int	true	"User ID"
//	@Success		204	{object}	User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		404	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/users/{id} [delete]
func (u *UserHandler) DeleteUser(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "DeleteUser functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// AssignRoleToUser assigns a role to a user
//
//	@Summary		Assign a role to a user
//	@Description	Assign a role to a user
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id		path		int		true	"User ID"
//	@Param			role	body		Role	true	"Role data"
//	@Success		200		{object}	User
//	@Failure		400		{object}	utils.ErrorResponse
//	@Failure		404		{object}	utils.ErrorResponse
//	@Failure		500		{object}	utils.ErrorResponse
//	@Router			/users/{id}/role [post]
func (u *UserHandler) AssignRoleToUser(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "AssignRoleToUser functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetUsersWithRoles retrieves all users with roles
//
//	@Summary		Get all users with roles
//	@Description	Retrieve all users with roles
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	[]User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/users/roles [get]
func (u *UserHandler) GetUsersWithRoles(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetUsersWithRoles functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// RemoveRoleFromUser removes a role from a user
//
//	@Summary		Remove a role from a user
//	@Description	Remove a role from a user
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Param			id		path		int		true	"User ID"
//	@Param			role	body		Role	true	"Role data"
//	@Success		200		{object}	User
//	@Failure		400		{object}	utils.ErrorResponse
//	@Failure		404		{object}	utils.ErrorResponse
//	@Failure		500		{object}	utils.ErrorResponse
//	@Router			/users/{id}/role [delete]
func (u *UserHandler) RemoveRoleFromUser(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "RemoveRoleFromUser functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}

// GetActiveUsers retrieves all active users
//
//	@Summary		Get all active users
//	@Description	Retrieve all active users
//	@Tags			users
//	@Accept			json
//	@Produce		json
//	@Success		200	{object}	[]User
//	@Failure		400	{object}	utils.ErrorResponse
//	@Failure		500	{object}	utils.ErrorResponse
//	@Router			/users/active [get]
func (u *UserHandler) GetActiveUsers(w http.ResponseWriter, r *http.Request) {
	errorResponse := utils.ErrorResponse{
		Code:    http.StatusNotImplemented, // 501
		Message: "GetActiveUsers functionality is not implemented.",
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	if err := json.NewEncoder(w).Encode(errorResponse); err != nil {
		http.Error(w, "An error occurred while processing the response", http.StatusInternalServerError)
	}
}